--- packages/vite-plugin-ssr-middleware/examples/README.md
@@ -1,1 +1,0 @@
-# examples
