--- packages/vite-plugin-ssr-middleware/examples/basic/README.md
@@ -1,5 +1,0 @@
-# basic
-
-```sh
-pnpm dev
-```
